src/requirements.txt
sentence-transformers==2.5.1
chromadb==0.4.22
openai==1.12.0
python-dotenv==1.0.1
rank-bm25==0.2.2
numpy==1.26.4